web/backend/requirements.txt
# Web后端依赖
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0

# 继承核心依赖
PyYAML>=6.0
Jinja2>=3.1.0
openai>=1.0.0
python-dotenv>=1.0.0

